#23 Added missing files and computed data for patter 25mm_10x7. Fixed some minor issues and some annoyances. Stabilized pocket detection.

+++ PoolSimulatorComponents/CameraAnalysis/objects_in_environment.py
@@ -6,7 +6,7 @@ import json
 @dataclass
 class TableSpec:
     name: str
-    playfield_mm: Tuple[int, int]
+    playfield_mm: Tuple[float, float]
     overall_mm: Optional[Tuple[int, int]] = None
     notes: str = ""
     cloth_profile: Optional[str] = None
